Add files via upload
This work is to complement our research. To be specific, It was published in land, a journal of MDPI Publishing, in the paper "Trade-Off and Coordination between Development and Ecological Protection of Urban Agglomerations.  along Rivers:  A Case Study of Urban Agglomerations in the Shandong Section of the Lower Yellow River (2.3. Intensity Analysis) supplement: strength analysis macros based on Microsoft Excel.